% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huxtableUtils.R
\name{defaultTableLayout}
\alias{defaultTableLayout}
\title{A tidy article theme for huxtables that works with google docs}
\usage{
defaultTableLayout(...)
}
\arguments{
\item{...}{a huxtable object}
}
\description{
Once copied and pasted table needs to have leading spaces removed in google docs
}
\examples{
defaultTableLayout(huxtable(dataframe \%>\% select("col 1 title"=col1))
}
\keyword{huxtable}
