% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omopVocab.R
\name{findConceptClasses}
\alias{findConceptClasses}
\title{list concept classes}
\usage{
findConceptClasses(searcher, class)
}
\arguments{
\item{searcher}{a searcher R6 object e.g. Searcher$new(omop)}

\item{class}{a concept class filter}
}
\description{
list concept classes
}
\examples{
omop = Omop$new(...)
omop$searcher$findConceptClasses("domain")
}
\keyword{omop}
