% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omopVocab.R
\name{findConcepts}
\alias{findConcepts}
\title{directly lookup concepts in db}
\usage{
findConcepts(searcher, term, vocabularyList = NA, domainList = NA,
  classList = NA, standardConcepts = TRUE)
}
\arguments{
\item{searcher}{a searcher R6 object e.g. Searcher$new(omop) or omop$searcher}

\item{term}{an input search string}

\item{vocabularyList}{a vocabulary filter (see findVocabularies)}

\item{domainList}{a domain filter (see findDomains)}

\item{classList}{a class filter (see findConceptClasses)}

\item{standardConcepts}{only return standard concepts}
}
\description{
directly lookup concepts in db
}
\examples{
omop = Omop$new(...)
omop$searcher$findConcepts(term)
}
\keyword{omop}
