% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omopVocab.R
\name{findDomains}
\alias{findDomains}
\title{list domains}
\usage{
findDomains(searcher, domain)
}
\arguments{
\item{searcher}{a searcher R6 object e.g. Searcher$new(omop)}

\item{vocabulary}{a vocabulary filter}
}
\description{
list domains
}
\examples{
omop = Omop$new(...)
omop$searcher$findDomains("domain")
}
\keyword{omop}
