% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omopVocab.R
\name{findVocabularies}
\alias{findVocabularies}
\title{list vocabularies}
\usage{
findVocabularies(searcher, vocab)
}
\arguments{
\item{searcher}{a searcher R6 object e.g. Searcher$new(omop)}

\item{vocabulary}{a vocabulary filter}
}
\description{
list vocabularies
}
\examples{
omop = Omop$new(...)
omop$searcher$findVocabularies(term)
}
\keyword{omop}
