% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOmop.R
\name{getOmop}
\alias{getOmop}
\title{Get a handle to the omop databse}
\usage{
getOmop(dbname = "omop", config = config::get(file =
  "~/Dropbox/db.yaml"))
}
\arguments{
\item{dbname}{String value of dbname defaults to "omop".}

\item{config}{Config object from yaml file.}
}
\description{
This function sets up connect to the omop databse.
}
\examples{
getOmop()
}
\keyword{omop}
