% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{meanAndConfidence}
\alias{meanAndConfidence}
\title{mean and confidence intervals using t-test}
\usage{
meanAndConfidence(x, f = twoDp, ...)
}
\arguments{
\item{x}{a numeric or vector of numerics}

\item{f}{a formatting function}
}
\description{
mean and confidence intervals using t-test
}
\examples{
df \%>\% vocabLookup(omop)
}
\keyword{format}
