% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotUtils.R
\name{narrowAndTall}
\alias{narrowAndTall}
\title{reorganise plot to minimise width (or maximise plotting width)}
\usage{
narrowAndTall()
}
\description{
reorganise plot to minimise width (or maximise plotting width)
}
\examples{
theme_set(themePhd())
}
\keyword{plot}
\keyword{theme}
