% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOmop.R
\docType{data}
\name{Omop}
\alias{Omop}
\title{Get a handle to the omop databse}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Omop
}
\arguments{
\item{dbname}{String value of dbname defaults to "omop".}

\item{config}{Config object from yaml file. (defaults to config::get(file="~/Dropbox/db.yaml"))}
}
\description{
This function sets up connect to the omop databse.
}
\examples{
omop = Omop$new(dbname,config)
}
\keyword{omop}
