% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotUtils.R
\name{savePubBig}
\alias{savePubBig}
\title{A standard publication shaped plotting}
\usage{
savePubBig(filename, plot = last_plot(), print_aspect_ratio = 1.1, ...)
}
\arguments{
\item{filename}{base of target filename (excuding extension).}

\item{plot}{a GGplot object or none}

\item{print_aspect_ratio}{svg aspect ratio}

\item{...}{passed to cowplot}
}
\description{
A standard publication shaped plotting
}
\examples{
savePubBig("the_filename",ggplot())
}
\keyword{axes}
