% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huxtableUtils.R
\name{saveTable}
\alias{saveTable}
\title{save labelled dataframe to html and tex file silently}
\usage{
saveTable(labelledDataframe, filename)
}
\arguments{
\item{filename}{file of desired output without extension.}

\item{labelledDataFrame}{e.g. dataframe %>% select("col 1 title"=col1)}
}
\description{
save labelled dataframe to html and tex file silently
}
\examples{
getOmop()
}
\keyword{omop}
