% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotUtils.R
\name{saveThesis}
\alias{saveThesis}
\title{A standard plot for publication}
\usage{
saveThesis(filename, maxWidth, maxHeight, plot = last_plot(),
  aspectRatio = NULL)
}
\arguments{
\item{filename}{base of target filename (excuding extension).}

\item{maxWidth}{maximum width in inches}

\item{maxHeight}{maximum height in inches}

\item{plot}{a GGplot object or none}

\item{aspectRatio}{defaults to maxWidth/maxHeight}
}
\description{
A standard plot for publication
}
\examples{
saveThesis("the_filename",maxWidth=4,maxHeight=4,plot=ggplot())
}
\keyword{plot}
