% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotUtils.R
\name{saveThesisFullPage}
\alias{saveThesisFullPage}
\title{A standard 6x8 plot size for a full page}
\usage{
saveThesisFullPage(filename, plot = last_plot(), aspectRatio = NULL)
}
\arguments{
\item{filename}{base of target filename (excuding extension).}

\item{plot}{a GGplot object or none}

\item{aspectRatio}{defaults to maxWidth/maxHeight}
}
\description{
A standard 6x8 plot size for a full page
}
\examples{
saveThesisFullPage("the_filename",ggplot())
}
\keyword{axes}
