% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotUtils.R
\name{saveThesisHalfPage}
\alias{saveThesisHalfPage}
\title{A standard max 6x4 plot size for a half page}
\usage{
saveThesisHalfPage(filename, plot = last_plot(), aspectRatio = NULL)
}
\arguments{
\item{filename}{base of target filename (excuding extension).}

\item{plot}{a GGplot object or none}

\item{aspectRatio}{defaults to maxWidth/maxHeight}
}
\description{
A standard max 6x4 plot size for a half page
}
\examples{
saveThesisFullPage("the_filename",ggplot())
}
\keyword{axes}
