% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotUtils.R
\name{saveThesisThirdPage}
\alias{saveThesisThirdPage}
\title{A standard max 6x3 plot size for a third page}
\usage{
saveThesisThirdPage(filename, ...)
}
\arguments{
\item{filename}{base of target filename (excuding extension).}

\item{...}{passed to saveThesis()}

\item{plot}{a GGplot object or none}
}
\description{
A standard max 6x3 plot size for a third page
}
\examples{
saveThesisFullPage("the_filename",ggplot())
}
\keyword{axes}
