% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omopVocab.R
\docType{data}
\name{Searcher}
\alias{Searcher}
\title{Get a searcher of the omop databse}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Searcher
}
\arguments{
\item{omop}{a R6 object of Omop class}
}
\description{
Get a searcher of the omop databse
}
\examples{
search = Searcher$new(omop)
}
\keyword{omop}
