% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotUtils.R
\name{themePhd}
\alias{themePhd}
\title{A phd ggplot2 theme}
\usage{
themePhd(base_size = 10, base_family = "Arial")
}
\arguments{
\item{base_size}{default 10}

\item{base_family}{default "Arial"}
}
\description{
A phd ggplot2 theme
}
\examples{
theme_set(themePhd())
}
\keyword{plot}
\keyword{theme}
