% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{twoDp}
\alias{twoDp}
\title{2 decimal place}
\usage{
twoDp(x, unit = "")
}
\arguments{
\item{x}{a numeric or vector of numerics}

\item{x}{a unit specification}
}
\description{
2 decimal place
}
\examples{
df \%>\% vocabLookup(omop)
}
\keyword{format}
