% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOmop.R
\name{vocabLookup}
\alias{vocabLookup}
\title{directly lookup concepts in db}
\usage{
vocabLookup(df, omop)
}
\arguments{
\item{df}{an input data frame}

\item{omop}{S6 connection to an omop database}
}
\description{
directly lookup concepts in db
}
\examples{
omop = Omop$new()
df \%>\% vocabLookup(omop)
df \%>\% omop$vocabLookup()
}
\keyword{omop}
