% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotUtils.R
\name{saveThesisSixthPage}
\alias{saveThesisSixthPage}
\title{A standard max 3x3 plot size for a page}
\usage{
saveThesisSixthPage(filename, ...)
}
\arguments{
\item{filename}{base of target filename (excuding extension).}

\item{...}{passed to saveThesis()}

\item{plot}{a GGplot object or none}
}
\description{
A standard max 3x3 plot size for a page
}
\examples{
saveThesisSixthPage("the_filename",ggplot())
}
\keyword{axes}
